tests: Fix time mocking on PyPy
Mock `datetime.datetime.now()` directly rather than the complete
`datetime` class, in order to fix compatibility with the pure Python
`datetime` stdlib implementation used on PyPy.  The original code failed
as it replaced the class with a `MagicMock` instance, while the new one
just overrides the method with a lambda.

Fixes #2708